I think everything works?

+++ views/Feed/Feed.jsx
@@ -99,15 +99,37 @@ export default class FeedView extends React.Component {
       });
   }
 
-  deletePost = async (postId) => {
-    return ApiClient.delete(`/posts/${postId}`, {authorized: true})
-      .then(() => {
-        this.updateState('deletePost', postId);
-      })
-      .catch(err => {
-        console.error(err);
-        alert("Error deleting post. Sorry about that!")
-      });
+  updatePost = async (postId, data, type) => {	
+    if (type === 'toggleLike') {	
+      const loggedInUser = this.props.navigation.getParam('loggedInUser');	
+      const addLike = data.usersLiked.some(user => user._id === loggedInUser._id);	
+
+      return ApiClient.post(`/posts/${postId}/like`, { addLike }, {authorized: true})	
+        .then(() => {	
+          this.updateState('updatePost', data);	
+        })	
+        .catch(err => {	
+          alert("Error liking post. Sorry about that!")	
+        });	
+    }
+    else if (type === 'deletePost') {	
+      return ApiClient.delete(`/posts/${postId}`, {authorized: true})	
+        .then(() => {	
+          this.updateState('deletePost', postId);	
+        })	
+        .catch(err => {	
+          alert("Error deleting post. Sorry about that!")	
+        });	
+    }
+
+    ApiClient.put(`/posts/${postId}`, _.pick(data, ['content', 'image']), {authorized: true})	
+    .then(() => {
+      this.loadData();
+    })
+    .catch(err => {
+      console.error(err);
+      alert("Error deleting post. Sorry about that!")
+    });
   }
 
   /**
@@ -196,7 +218,7 @@ export default class FeedView extends React.Component {
         showUserProfile={this.showUserProfile}
         showFullDate={false}
         navigation={this.props.navigation}
-        deletePost={this.deletePost}
+        updatePost={this.updatePost}
       />
     );
   }
